reskit: Nice selection box of goodies for hobnobbing with NHP model results

'Wafer' goodbye to crumbly code, and dunk a sturdy reskit function in your data cuppa instead. We've clustered the cream of our results code here - it's a cracker!

Author

Maintainer: Fran Barton francis.barton@nhs.net (ORCID)

Authors:

Fran Barton francis.barton@nhs.net (ORCID)

Other contributors:

Matt Dray matt.dray@nhs.net [contributor]

Rhian Davies rhian.davies25@nhs.net [contributor]